+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/NavigationConstants.java
@@ -80,8 +80,10 @@ public final class NavigationConstants {
      * Maximum number of meters the user can travel away from step before the
      * {@link OffRouteListener}'s called.
      *
+     * @deprecated has no effect and will be removed in a future release.
      * @since 0.2.0
      */
+    @Deprecated
     static final double MAXIMUM_DISTANCE_BEFORE_OFF_ROUTE = 20;
 
     /**
@@ -119,7 +121,17 @@ public final class NavigationConstants {
      */
     public static final double METERS_REMAINING_TILL_ARRIVAL = 40;
 
-    public static final double MINIMUM_BACKUP_DISTANCE_FOR_OFF_ROUTE = 50;
+    /**
+     * Minimum distance in meters that the user must travel in the wrong direction before the
+     * off-route logic recognizes the user is moving away from upcoming maneuver
+     */
+    public static final double OFF_ROUTE_MINIMUM_DISTANCE_METERS_BEFORE_WRONG_DIRECTION = 50;
+
+    /**
+     * Minimum distance in meters that the user must travel in the correct direction before the
+     * off-route logic recognizes the user is back on the right direction
+     */
+    public static final double OFF_ROUTE_MINIMUM_DISTANCE_METERS_BEFORE_RIGHT_DIRECTION = 20;
 
     public static final double MINIMUM_DISTANCE_BEFORE_REROUTING = 50;